Data can now be sent with item actions
This small change allows the client to send any kind of data along with
an "action" request on an item.

This can extend and simplify interactivity functionality for items. A
usage example:
A notepad item which you can write on. Instead of using additional
netstreams to network text to the server, the write function now opens a
VGUI panel on the client, and when the player decides to save their
changes, the panel starts an "invAct" netstream with this additional
data attached to it. Serverside, the item can now decide what to do with
this data.

With a few utility functions, this can be made as simple as calling
something like item:callOnServer(...) in a VGUI panel.